# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /institutions:
    get:
      operationId: listInstitutions
      tags: [institutions]
      parameters:
        - in: query
          name: mine
          required: false
          description: When 1, account_manager admins see only institutions assigned to them
          schema: { type: string, enum: ["1"] }
      responses:
        "200":
          description: Institutions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Institution" }
    post:
      operationId: createInstitution
      tags: [institutions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInstitution" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Institution" }
  /institutions/{id}:
    get:
      operationId: getInstitution
      tags: [institutions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Institution detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionDetail" }
  /institutions/{id}/students:
    get:
      operationId: listInstitutionStudents
      tags: [institutions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Students at institution with placement status
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InstitutionStudent" }
  /institutions/{id}/students/{candidateId}/verify:
    post:
      operationId: verifyInstitutionStudent
      tags: [institutions]
      summary: Mark a candidate as a verified student of this institution.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: candidateId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyStudentResponse" }
  /institutions/{id}/students/{candidateId}/unverify:
    post:
      operationId: unverifyInstitutionStudent
      tags: [institutions]
      summary: Revoke verification for a previously verified student.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: candidateId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Unverified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    VerifyStudentResponse:
      type: object
      properties:
        ok: { type: boolean }
        verifiedAt: { type: string, format: date-time }
      required: [ok, verifiedAt]
...
    Employer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        tagline: { type: string }
        description: { type: string }
        industry: { type: string }
        location: { type: string }
        logoUrl: { type: string }
        coverUrl: { type: string }
        websiteUrl: { type: string }
        size:
          type: string
          enum: [startup, small, midsize, large, enterprise]
        verified: { type: boolean }
        openJobs: { type: integer }
        createdAt: { type: string, format: date-time }
        accountManagerId:
          type: integer
          nullable: true
          description: Owning account-manager admin user id (admin-only field)
        accountManagerName:
          type: string
          nullable: true
          description: Owning account-manager display name (admin-only field)
      required:
        - id
        - name
        - tagline
        - description
        - industry
        - location
        - logoUrl
        - coverUrl
        - websiteUrl
        - size
        - verified
        - openJobs
        - createdAt
...
    Institution:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type:
          type: string
          enum: [university, college, bootcamp, vocational, other]
        location: { type: string }
        logoUrl: { type: string }
        websiteUrl: { type: string }
        studentCount: { type: integer }
        placementRate: { type: number }
        createdAt: { type: string, format: date-time }
        accountManagerId:
          type: integer
          nullable: true
          description: Owning account-manager admin user id (admin-only field)
        accountManagerName:
          type: string
          nullable: true
          description: Owning account-manager display name (admin-only field)
      required:
        - id
        - name
        - type
        - location
        - logoUrl
        - websiteUrl
        - studentCount
        - placementRate
        - createdAt
    InstitutionDetail:
      allOf:
        - $ref: "#/components/schemas/Institution"
        - type: object
          properties:
            description: { type: string }
            partnerEmployers:
              type: array
              items: { $ref: "#/components/schemas/Employer" }
          required: [description, partnerEmployers]
    InstitutionStudent:
      type: object
      properties:
        candidateId: { type: integer }
        fullName: { type: string }
        avatarUrl: { type: string }
        headline: { type: string }
        talentScore: { type: integer }
        readinessScore: { type: integer }
        status:
          type: string
          enum: [active, applying, interviewing, hired, graduated]
        currentEmployerName: { type: string, nullable: true }
        applicationsCount: { type: integer }
        isPrimaryAffiliation:
          type: boolean
          description: True if this institution is the student's primary affiliation; false if it's a secondary (e.g. transfer or bootcamp) link.
        isVerified:
          type: boolean
          description: True if this institution has explicitly verified the candidate as a real student. Unverified students are excluded from tracking metrics.
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        verifiedByName:
          type: string
          nullable: true
          description: Name of the institution staff member who verified this student.
      required:
        - candidateId
        - fullName
        - avatarUrl
        - headline
        - talentScore
        - readinessScore
        - status
        - applicationsCount
        - isPrimaryAffiliation
        - isVerified
    CreateInstitution:
      type: object
      properties:
        name: { type: string }
        type:
          type: string
          enum: [university, college, bootcamp, vocational, other]
        location: { type: string }
        logoUrl: { type: string }
        websiteUrl: { type: string }
        description: { type: string }
      required: [name, type, location, logoUrl, websiteUrl, description]
...